chore: add CastTxSender (#12834)
